+++ Lineaire_Algebra/TranslationMatrix.h
@@ -4,11 +4,11 @@
 class TranslationMatrix : public Matrix
 {
 	public:
-		TranslationMatrix(int n, int m, float x, float y, float z);
+		TranslationMatrix(int n, int m, double x, double y, double z);
 		~TranslationMatrix() override;
 };
 
-TranslationMatrix::TranslationMatrix(int n, int m, float x, float y, float z) : Matrix(n, m)
+TranslationMatrix::TranslationMatrix(int n, int m, double x, double y, double z) : Matrix(n, m)
 {
 	for (int i = 0; i < n; i++)
 		(*this)(i, i) = 1;
